refactor compatibility module
